+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### New features
 
+- New `\WizaplaceFrontBundle\Service\FavoriteService` which decorates the SDK's `FavoriteService` with a request-scoped cache
+
 ### Bugfixes
 
 </details>