+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### New features
 
 - New `\WizaplaceFrontBundle\Service\FavoriteService` which decorates the SDK's `FavoriteService` with a request-scoped cache
+- Translate authentication flash messages
 
 ### Bugfixes